# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_request.PayrollSettingByKeyRequest:
    properties:
      config_key:
        maxLength: 255
        minLength: 3
        type: string
      value:
        type: number
    required:
    - config_key
    - value
    type: object
  github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_request.PayrollSettingIdsRequest:
    properties:
      ids:
        items:
          type: string
        minItems: 1
        type: array
    required:
    - ids
    type: object
  github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_request.PayrollSettingRequest:
    properties:
      config_name:
        maxLength: 255
        minLength: 3
        type: string
      value:
        type: number
    required:
    - config_name
    - value
    type: object
  github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_request.PayrollSettingUpdateBulkRequest:
    properties:
      settings:
        items:
          $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_request.PayrollSettingByKeyRequest'
        minItems: 1
        type: array
    required:
    - settings
    type: object
...
  github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response.PayrollSettingDeleteResponse:
    properties:
      deleted_count:
        type: integer
      skipped_count:
        type: integer
      total:
        type: integer
    type: object
  github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response.PayrollSettingResponse:
    properties:
      config_key:
        type: string
      config_name:
        type: string
      created_at:
        type: string
      id:
        type: string
      updated_at:
        type: string
      value:
        type: number
    type: object
...
  github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.ErrorCode:
    enum:
    - BAD_REQUEST
    - VALIDATION_ERROR
    - AUTH_REQUIRED
    - AUTH_INVALID_CREDENTIALS
    - NOT_FOUND
    - FORBIDDEN
    - INTERNAL_SERVER_ERROR
    - DATABASE_ERROR
    - TOO_MANY_REQUESTS
    type: string
    x-enum-varnames:
    - BAD_REQUEST
    - VALIDATION_ERROR
    - AUTH_REQUIRED
    - AUTH_INVALID_CREDENTIALS
    - NOT_FOUND
    - FORBIDDEN
    - INTERNAL_SERVER_ERROR
    - DATABASE_ERROR
    - TOO_MANY_REQUESTS
  github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.ErrorSchema:
    properties:
      code:
        $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.ErrorCode'
      details: {}
      message:
        type: string
    type: object
  github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.ErrorSchema'
      success:
        type: boolean
    type: object
...
  ? github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-array_github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_PayrollSettingResponse
  : properties:
      data:
        items:
          $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response.PayrollSettingResponse'
        type: array
      error:
        $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.ErrorSchema'
      success:
        type: boolean
    type: object
...
  ? github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_PayrollSettingDeleteResponse
  : properties:
      data:
        $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response.PayrollSettingDeleteResponse'
      error:
        $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.ErrorSchema'
      success:
        type: boolean
    type: object
  ? github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_PayrollSettingResponse
  : properties:
      data:
        $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response.PayrollSettingResponse'
      error:
        $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.ErrorSchema'
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: afifudin23@example.com
    name: Afifudin
  description: REST API untuk manajemen absensi, karyawan, dan penggajian King Royal.
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: Absensi King Royal API
  version: "1.0"
paths:
...
  /payroll-settings:
    delete:
      consumes:
      - application/json
      description: |-
        Menghapus satu atau beberapa komponen konfigurasi penggajian berdasarkan daftar ID.

        **Request Body:**
        - `ids` (array of UUID, required, min 1): Daftar ID komponen yang akan dihapus

        **Response 200:** `total` (jumlah yang dihapus), `deleted_count`, `skipped_count`
      parameters:
      - description: Daftar ID yang akan dihapus
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_request.PayrollSettingIdsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_PayrollSettingDeleteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
      security:
      - BearerAuth: []
      summary: Hapus pengaturan payroll
      tags:
      - Payroll Settings
    get:
      description: |-
        Mendapatkan semua komponen konfigurasi penggajian yang tersimpan.

        **Response 200:** Array konfigurasi (id, config_name, config_key, value, tanggal dibuat/diupdate)

        **Contoh komponen:** Tarif lembur, potongan absensi, pajak penghasilan, tunjangan default, dll
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-array_github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_PayrollSettingResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
      security:
      - BearerAuth: []
      summary: Daftar pengaturan payroll
      tags:
      - Payroll Settings
    post:
      consumes:
      - application/json
      description: |-
        Menambahkan komponen konfigurasi penggajian baru.

        **Request Body:**
        - `config_name` (string, required, min 3 karakter): Nama komponen (contoh: "Tarif Lembur Per Jam")
        - `value` (number, required): Nilai komponen (contoh: 50000)

        **Behavior:**
        - `config_key` dibuat otomatis dari `config_name` (snake_case)
        - Nama komponen harus unik, jika sudah ada akan mengembalikan 400

        **Response 201:** Data komponen yang baru dibuat (id, config_name, config_key, value)
      parameters:
      - description: Data komponen penggajian
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_request.PayrollSettingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_PayrollSettingResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
      security:
      - BearerAuth: []
      summary: Buat pengaturan payroll
      tags:
      - Payroll Settings
  /payroll-settings/{payroll_id}:
    patch:
      consumes:
      - application/json
      description: |-
        Memperbarui satu komponen konfigurasi penggajian berdasarkan ID.

        **Request Body:**
        - `config_name` (string, required): Nama komponen baru
        - `value` (number, required): Nilai komponen baru

        **Response 201:** Data komponen setelah diupdate
      parameters:
      - description: ID pengaturan payroll
        in: path
        name: payroll_id
        required: true
        type: string
      - description: Data yang diperbarui
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_request.PayrollSettingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_PayrollSettingResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
      security:
      - BearerAuth: []
      summary: Update pengaturan payroll
      tags:
      - Payroll Settings
  /payroll-settings/bulk:
    put:
      consumes:
      - application/json
      description: |-
        Memperbarui beberapa komponen konfigurasi penggajian sekaligus dalam satu request.

        **Request Body:**
        - `settings` (array, required, min 1): Daftar komponen yang akan diupdate
        - `config_key` (string, required): Key komponen yang akan diupdate (harus sudah ada)
        - `value` (number, required): Nilai baru

        **Response 201:** Array seluruh komponen setelah diupdate
      parameters:
      - description: Daftar komponen yang diperbarui
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_request.PayrollSettingUpdateBulkRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-array_github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_PayrollSettingResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
      security:
      - BearerAuth: []
      summary: Bulk update pengaturan payroll
      tags:
      - Payroll Settings
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token JWT dengan format: **Bearer &lt;token&gt;** — contoh:
      `Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...`'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...